Backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.28
requests==2.31.0
textblob==0.17.1
scikit-learn==1.3.2
pandas==2.1.4
xgboost==2.0.3
fredapi==0.5.1
python-dotenv==1.0.0
